Update playersite on saving a channel (datarhei/restreamer#362)
In case the playersite is enabled and the metadata of a channel e.g. its name
changes, the playersite needs also to be updated in order to reflect this
change.

As a consequence the playersite has to be enabled explicitely otherwise it will
get enabled without the user wanting it.