Add changelog notification dialog
Expecting the CHANGELOG.md in the UI webroot. The markdown is
parsed in order to find the sections with the versions. A section
for a version has to start with "### " followed by the version
(either with the v prefix or not). Only the sections for the
relevant versions from the previous to the current will be displayed.